# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Language Learning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerWithEmail
      tags: [auth]
      summary: Register with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRegisterRequest"
      responses:
        "200":
          description: Registration result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthActionResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/login-email:
    post:
      operationId: loginWithEmail
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailLoginRequest"
      responses:
        "200":
          description: Login result with session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthActionResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/verify-email:
    get:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email address via token
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthActionResult"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Always returns ok (to avoid email enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthActionResult"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Reset result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthActionResult"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
...
components:
  schemas:
...
    EmailRegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required: [email, password, name]
    EmailLoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
      required: [email]
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required: [token, password]
    AuthActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    AuthError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...